Moving some types and classes from PlaySupport to play
- updated tests and readme accordingly
