+++ README.md
@@ -24,12 +24,12 @@ libraryDependencies += "com.autoscout24" %% "toguru-scala-client" % "(see versio
 
 To toggle code with this client, you need to perform the following steps.
 
-Create a toguru client - e.g. in the Guice module of your Play
+Create a Toguru client - e.g. in the Guice module of your Play
 application so that it can be injected wherever you need it. For this,
 you need to define
-* how toggling relevant client information (e.g. client id and user agent)
-  can be extracted from a play request, and
-* the toguru server where the toggle activation conditions can be fetched from.
+* how toggling-relevant client information (e.g. client id and user agent)
+  can be extracted from a Play request, and
+* the Toguru server where the toggle activation conditions can be fetched from.
 
 ```scala
 import play.api.mvc._
@@ -45,7 +45,7 @@ val toguruClient = PlaySupport.toguruClient(client, "http://localhost:9000")
 
 Define a toggle using a toggle id and a default activation condition, e.g.
 in an object that contains all your toggles. The activation condition
-will be fetched from the toguru server based on the toggle id. The
+will be fetched from the Toguru server based on the toggle id. The
 fallback is used if the server can't be reached or the toggle id isn't
 in the server's toggle state response.
 
@@ -55,7 +55,7 @@ val toggle = Toggle("my-toggle", default = Condition.Off)
 
 
 Now, you can determine the state of your toggle based on the client info
-and the toggle activation condition from the toguru server:
+and the toggle activation condition from the Toguru server:
 
 ```scala
 implicit val toggling = toguruClient(request)
@@ -66,14 +66,13 @@ else
   Ok("Toggle is off")
 ```
 
-The toguru Scala client offers several convenience utilities, however.
-To begin with, you can create a toggled controller based on the toguru
+The Toguru Scala client offers several convenience utilities, however.
+To begin with, you can create a toggled controller based on the Toguru
 client that your controllers can extend from:
 
 ```scala
 import play.api.mvc._
-import toguru.play.PlaySupport
-import toguru.play.PlaySupport.PlayToguruClient
+import toguru.play._
 
 abstract class ToggledController(toguru: PlayToguruClient) extends Controller {
   val ToggledAction = PlaySupport.ToggledAction(toguru)
@@ -81,7 +80,7 @@ abstract class ToggledController(toguru: PlayToguruClient) extends Controller {
 ```
 
 
-Now you can define your controller with toggled actions and use the 
+Now you can define your controller with toggled actions and use the
 toggle defined earlier to control which code gets executed:
 
 ```scala
@@ -109,7 +108,7 @@ class MyRequest[A](toguru: PlayToguruClient, request : Request[A])
 }
 ```
 
-Alternatively, you can alwazs create the toggle information yourself:
+Alternatively, you can always create the toggle information yourself:
 
 ```scala
 class MyControllerWithOwnTogglingInfo @Inject()(toguru: PlayToguruClient) extends Controller {
@@ -138,7 +137,7 @@ val toguruClient = PlaySupport.toguruClient(client, TestActivations(toggle -> Co
 val controller = new MyController(toguruClient)
 ```
 
-see the [PlaySupport spec](src/test/scala/toguru/play/PlaySupportSpec) for 
+see the [PlaySupport spec](src/test/scala/toguru/play/PlaySupportSpec.scala) for
 full usage examples in Scala Play.
 
 ## Copyright
